More test stubs for PublishPage and RevertPage...
