+++ app/Providers/AppServiceProvider.php
@@ -24,7 +24,7 @@ class AppServiceProvider extends ServiceProvider
 	{
 		Schema::defaultStringLength(191); // Fix for the webfarm running older MySQL
 
-        Validator::extend('parent_is_published', function( $attr, $value ) { return PublishPage::parentIsPublished($value); });
+        Validator::extend('parent_is_published', function( $attr, $value ) { return PublishPage::canBePublished($value); });
         Validator::extend('unique_site_path', function($attr, $value, $parameters, $validator) {
             $host = isset($parameters[0]) ? $parameters[0] : null;
             return (new UniqueSitePathRule($host))->passes($attr, $value);
